+++ .gitignore
@@ -25,5 +25,3 @@ testfile
 
 
 !/.github\ISSUE_TEMPLATE\config.yml
-
-test-results
