+++ tests/pytest.ini
@@ -1,7 +1,7 @@
 [pytest]
 asyncio_mode = auto
 timeout = 20
-addopts = --alluredir allure-results --tracing retain-on-failure
+addopts = --alluredir=allure-results --tracing=retain-on-failure --screenshot=only-on-failure
 log_cli = 1
 log_cli_level = DEBUG
 log_cli_format = %(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)
